Add files via upload
Create a simple web structure where you are asking user for the task and once user click on the button, that task will be displayed on the screen in a list format. Along with that, add a button to delete a task as well. Marking will be done on the basis of your JavaScript knowledge used over here.